move driver config functions
