+++ src/dmx/driver.h
@@ -41,6 +41,44 @@ extern "C" {
 #define DMX_CHECK(a, err_code, format, ...) \
   ESP_RETURN_ON_FALSE(a, err_code, TAG, format, ##__VA_ARGS__)
 
+enum dmx_interrupt_mask_t {
+  DMX_INTR_RX_FIFO_OVERFLOW = UART_INTR_RXFIFO_OVF,
+  DMX_INTR_RX_FRAMING_ERR = UART_INTR_PARITY_ERR | UART_INTR_FRAM_ERR,
+  DMX_INTR_RX_ERR = DMX_INTR_RX_FIFO_OVERFLOW | DMX_INTR_RX_FRAMING_ERR,
+
+  DMX_INTR_RX_BREAK = UART_INTR_BRK_DET,
+  DMX_INTR_RX_DATA = UART_INTR_RXFIFO_FULL,
+  DMX_INTR_RX_ALL = DMX_INTR_RX_DATA | DMX_INTR_RX_BREAK | DMX_INTR_RX_ERR,
+
+  DMX_INTR_TX_DATA = UART_INTR_TXFIFO_EMPTY,
+  DMX_INTR_TX_DONE = UART_INTR_TX_DONE,
+  DMX_INTR_TX_ALL = DMX_INTR_TX_DATA | DMX_INTR_TX_DONE,
+
+  DMX_ALL_INTR_MASK = -1
+};
+
+enum rdm_packet_timing_t {
+  RDM_DISCOVERY_NO_RESPONSE_PACKET_SPACING = 5800,
+  RDM_REQUEST_NO_RESPONSE_PACKET_SPACING = 3000,
+  RDM_BROADCAST_PACKET_SPACING = 176,
+  RDM_RESPOND_TO_REQUEST_PACKET_SPACING = 176,
+
+  RDM_CONTROLLER_RESPONSE_LOST_TIMEOUT = 2800,
+  RDM_RESPONDER_RESPONSE_LOST_TIMEOUT = 2000
+};
+
+enum rdm_packet_type_t {
+  RDM_PACKET_TYPE_NON_RDM = 0,
+  RDM_PACKET_TYPE_DISCOVERY = BIT(0),
+  RDM_PACKET_TYPE_DISCOVERY_RESPONSE = BIT(1),
+  RDM_PACKET_TYPE_REQUEST = BIT(2),
+  RDM_PACKET_TYPE_RESPONSE = BIT(3),
+  RDM_PACKET_TYPE_BROADCAST = BIT(4),
+
+  RDM_PACKET_TYPE_EARLY_TIMEOUT =
+      (RDM_PACKET_TYPE_REQUEST | RDM_PACKET_TYPE_DISCOVERY)
+};
+
 /** @brief The DMX driver object used to handle reading and writing DMX data on
  * the UART port. It storese all the information needed to run and analyze DMX
  * and RDM.*/
